+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.2:
 
 * #2 Support for authentication filters
+* #3 Switch to ServiceLoader-based polymorphic configs for AuthenticatorFactory
 
 ## 0.1: